refactor: extract projectIterateLean to iterate-lean.mts; gate threads/checks/instructions/triggers
Splits the file to stay under the 200-line oxlint limit.
Also gates threads, fix.checks, instructions, and escalate.triggers
on non-empty — consistent with the other lean-array fields.
